# docs/swagger.yaml
basePath: /
definitions:
  apis.AcceptResetPassword:
    properties:
      key:
        type: string
      newPassword:
        type: string
    type: object
  apis.ApiResponse:
    properties:
      code:
        type: string
      data:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
...
  apis.CreateTeamResponse:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/models.Team'
      message:
        type: string
    type: object
...
  apis.LisTeamResponse:
    properties:
      code:
        type: string
      data:
        items:
          $ref: '#/definitions/models.Team'
        type: array
      message:
        type: string
    type: object
  apis.LoginPayload:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
...
  apis.ResetPassword:
    properties:
      email:
        type: string
    type: object
...
  apis.TeamPayload:
    properties:
      description:
        type: string
      name:
        type: string
    required:
    - description
    - name
    type: object
  apis.UpdateTeamResponse:
    properties:
      code:
        type: string
      data:
        $ref: '#/definitions/models.Team'
      message:
        type: string
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Bot:
    properties:
      desc:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.BotToken:
    properties:
      bot:
        $ref: '#/definitions/models.Bot'
      bot_id:
        type: integer
      created_at:
        type: string
      device:
        $ref: '#/definitions/models.Device'
      device_id:
        type: integer
      id:
        type: integer
      last_login:
        type: string
      team:
        $ref: '#/definitions/models.Team'
      team_id:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: integer
    type: object
  models.Device:
    properties:
      fingerprintID:
        type: string
      hostname:
        type: string
      id:
        type: integer
      macs:
        items:
          type: string
        type: array
      platform:
        type: string
    required:
    - hostname
    - macs
    - platform
    type: object
  models.Quota:
    properties:
      bot:
        $ref: '#/definitions/models.Bot'
      botID:
        type: integer
      count:
        type: integer
      id:
        type: integer
      limit:
        type: integer
      team:
        $ref: '#/definitions/models.Team'
      teamID:
        type: integer
    type: object
  models.Team:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      quotas:
        items:
          $ref: '#/definitions/models.Quota'
        type: array
      users:
        items:
          $ref: '#/definitions/models.User'
        type: array
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      lastReset:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
      suspended:
        type: boolean
      teams:
        items:
          $ref: '#/definitions/models.Team'
        type: array
      token:
        items:
          $ref: '#/definitions/models.BotToken'
        type: array
      updatedAt:
        type: string
      username:
        type: string
      verified:
        type: boolean
    required:
    - email
    - name
    - password
    - username
    type: object
host: localhost:5000
info:
  contact: {}
  title: Authentification API documentation
  version: 1.0.0
paths:
  /accept_reset_pwd:
    post:
      consumes:
      - application/json
      description: Reset Password Request pertama
      parameters:
      - description: reset
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/apis.AcceptResetPassword'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.ApiResponse'
      summary: Reset Password
      tags:
      - Users
...
  /login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/apis.LoginPayload'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.ApiResponse'
      summary: Login user
      tags:
      - Users
...
  /register:
    post:
      consumes:
      - application/json
      description: Create new user
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Create new user based on paramters
      tags:
      - Users
  /reset_pwd:
    post:
      consumes:
      - application/json
      description: Reset Password Request pertama
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/apis.ResetPassword'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.ApiResponse'
      summary: Reset Password
      tags:
      - Users
  /search_user:
    get:
      consumes:
      - application/json
      description: List User
      parameters:
      - description: reset
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/apis.AcceptResetPassword'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.ApiResponse'
      summary: List User
      tags:
      - Users
  /team:
    delete:
      description: Remove User Dari Team
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.ApiResponse'
      summary: Remove User Dari Team
      tags:
      - Teams
    get:
      consumes:
      - application/json
      description: Remove User Dari Team
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.LisTeamResponse'
      summary: Remove User Dari Team
      tags:
      - Teams
    post:
      consumes:
      - application/json
      description: create team
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/apis.LoginPayload'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.CreateTeamResponse'
      summary: Untuk create Team
      tags:
      - Teams
    put:
      consumes:
      - application/json
      description: Remove User Dari Team
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/apis.TeamPayload'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.UpdateTeamResponse'
      summary: Remove User Dari Team
      tags:
      - Teams
...
swagger: "2.0"